Fix class method aliasing bug in CoffeeScript.
In a class definition, aliasing a method like this:

    class Node
        del: @delete

Results in generated JS like this:

    var Node;
    Node = (function() {
      function Node() {}
      Node.prototype.del = Node["delete"];
      return Node;
    })();

Notice how `Node.prototype.del` is getting aliased to the **static**
`Node.delete`, when we want it aliased to the **prototype** `Node.delete`.

Fix by using the `::` prototype operator:

    class Node
        del: @::delete

It's surprising to me that our PropertyContainer's `@getter` and `@setter`
worked up until now. We lucked out because the actual implementations of
`Object.prototype.__defineGetter__` and `__defineSetter__` aren't bound to
the original instances but rather rely on the dynamic value of `this`, so
aliasing to the static `PropertyContainer.__defineGetter__` etc. worked.