Delete node relationships only if forced to.
This prevents oversight mistakes. E.g. if you delete a feed story, it's easy
to manually delete the relationships for what the story relates to, but it's
easy to forget to check for comments, and deleting this feed story would thus
break that comment node (it will be missing a target relationship).