API: throw in getNodeById() for non-existent IDs.
If we're doing a lookup explicitly by ID, we expect the ID to exist. It's an
exception if the ID doesn't exist; we should fail fast.

Makes sense from a REST perspective, too: 404 means client error. In the case
of this library, this was the only 404 case for which we weren't throwing.