Hotfix: freeze Streamline dependency.
We were using Streamline 0.1.35 locally, and 0.1.37 introduced a regression
with futures. It's been quickly fixed in 0.1.39, but I'm not sure yet if
0.1.37 is a safe upgrade yet: it awesomely transforms error stack traces to
synchronous stacks instead of async ones, but it might not work as expected if
the code doesn't explicitly require the flows library, which we don't.
Investigating that, but for now, freezing the dependency on 0.1.35 to protect
against the regression.