Loosen Streamline dependency semver spec.
We still need this to be a formal dependency since compiled Streamline files
still depend on a (minimal) Streamline "runtime".

We seem to both compile and run fine with the current Streamline 0.8.1, and
Bruno has been great about not introducing backwards-incompatible/breaking
changes to the *runtime*, so this lets us stay on top of bug fixes etc.