Create node/relationship index with configuration
- A 2nd parameter, which accepts an index config object, is added to
both createNodeIndex and createRelationshipIndex methods.  Their
overload helper methods are also added.
- 4 test cases are added with two different configs; fulltext with
lowercase support and fulltext with no lowercase support for both node
indexes and relationship indexes.  4 other test cases are added for
indexing nodes and relationships to custom configured indexes.  Total
of 8 tests are added.  Existing tests are not modified.
- Indexed nodes and relationships are now deleted after tests are
completed in order to prevent a confilct with previous and next tests.
(This issue was a todo of one of the contributors)
- There was no such method called queryRelationshipIndex implemented
which is required for retrieving relations from fulltext indexes, so I
implemented one.  It was also a todo of another contributor so I
deleted the following todo comment now that method is implemented.  I
have not added an extra test for this method because it is used in two
other tests I've implemented to retrieve indexed relationships